fix: sign-compare in trend hysteresis

+++ KEITHLEY_2000_LCD/Core/Src/main.c
@@ -1840,10 +1840,10 @@ static void trend_draw_column(uint16_t column, bool erase_previous)
     if (!s_trend_full_repaint &&
         occupied == trend_drawn_occupied(column) &&
         (!occupied ||
-         ((s_drawn_trend_y0[s_render_page][column] > y0
+         ((uint8_t)(s_drawn_trend_y0[s_render_page][column] > y0
                ? s_drawn_trend_y0[s_render_page][column] - y0
                : y0 - s_drawn_trend_y0[s_render_page][column]) < 2u &&
-          (s_drawn_trend_y1[s_render_page][column] > y1
+          (uint8_t)(s_drawn_trend_y1[s_render_page][column] > y1
                ? s_drawn_trend_y1[s_render_page][column] - y1
                : y1 - s_drawn_trend_y1[s_render_page][column]) < 2u)))
     {
